+++ x/feemarket/keeper/keeper.go
@@ -20,7 +20,6 @@ import (
 
 	corestoretypes "cosmossdk.io/core/store"
 	"cosmossdk.io/log"
-	"cosmossdk.io/store/prefix"
 	storetypes "cosmossdk.io/store/types"
 	"github.com/cosmos/cosmos-sdk/codec"
 	sdk "github.com/cosmos/cosmos-sdk/types"
@@ -37,8 +36,7 @@ type Keeper struct {
 	cdc          codec.BinaryCodec
 	storeService corestoretypes.KVStoreService
 	// Store key required for the Fee Market Prefix KVStore.
-	storeKey     storetypes.StoreKey
-	transientKey storetypes.StoreKey
+	storeKey storetypes.StoreKey
 	// the address capable of executing a MsgUpdateParams message. Typically, this should be the x/gov module account.
 	authority sdk.AccAddress
 }
@@ -48,7 +46,7 @@ func NewKeeper(
 	cdc codec.BinaryCodec,
 	storeService corestoretypes.KVStoreService,
 	authority sdk.AccAddress,
-	storeKey, transientKey storetypes.StoreKey,
+	storeKey storetypes.StoreKey,
 ) Keeper {
 	// ensure authority account is correctly formatted
 	if err := sdk.VerifyAddressFormat(authority); err != nil {
@@ -60,7 +58,6 @@ func NewKeeper(
 		storeService: storeService,
 		storeKey:     storeKey,
 		authority:    authority,
-		transientKey: transientKey,
 	}
 }
 
@@ -92,42 +89,6 @@ func (k Keeper) GetBlockGasWanted(ctx sdk.Context) uint64 {
 	}
 
 	return sdk.BigEndianToUint64(bz)
-}
-
-func (k Keeper) SumTransientGasWanted(ctx sdk.Context) uint64 {
-	store := prefix.NewStore(ctx.TransientStore(k.transientKey), types.KeyPrefixTransientBlockGasWanted)
-	it := store.Iterator(nil, nil)
-	defer it.Close()
-
-	var result uint64
-	for ; it.Valid(); it.Next() {
-		result += sdk.BigEndianToUint64(it.Value())
-	}
-	return result
-}
-
-// GetTransientGasWanted returns the gas wanted in the current block from transient store.
-func (k Keeper) GetTransientGasWanted(ctx sdk.Context) uint64 {
-	store := ctx.TransientStore(k.transientKey)
-	bz := store.Get(types.TransientBlockGasWantedKey(ctx.TxIndex()))
-	if len(bz) == 0 {
-		return 0
-	}
-	return sdk.BigEndianToUint64(bz)
-}
-
-// SetTransientBlockGasWanted sets the block gas wanted to the transient store.
-func (k Keeper) SetTransientBlockGasWanted(ctx sdk.Context, gasWanted uint64) {
-	store := ctx.TransientStore(k.transientKey)
-	gasBz := sdk.Uint64ToBigEndian(gasWanted)
-	store.Set(types.TransientBlockGasWantedKey(ctx.TxIndex()), gasBz)
-}
-
-// AddTransientGasWanted adds the cumulative gas wanted in the transient store
-func (k Keeper) AddTransientGasWanted(ctx sdk.Context, gasWanted uint64) (uint64, error) {
-	result := k.GetTransientGasWanted(ctx) + gasWanted
-	k.SetTransientBlockGasWanted(ctx, result)
-	return result, nil
 }
 
 // GetBaseFeeV1 get the base fee from v1 version of states.